Refactor: refactor workbook api directory
