+++ src/components/workbook-detail-drawer.tsx
@@ -11,8 +11,8 @@ import { IWorkbookBase } from '@/model/i-workbook';
 import { DialogDescription, DialogTitle } from '@radix-ui/react-dialog';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { useParams } from 'next/navigation';
-import { createWorkbook } from '@/app/(after-login)/(nav)/workbook/workbook-api';
 import { updateWorkbookDetail } from '@/app/(after-login)/(top)/workbook-detail/[workbookId]/workbook-detail-api';
+import { createWorkbook } from '@/api/workbook-api';
 
 const WorkbookDetailDrawer = ({
   isOpen,
